requirements.txt
# Using InfluxDB version 1.x (not 2.x)
influxdb

# GB Database is retrieved via WebDAV (e.g. NextCloud server)
webdavclient3

# Only required when webhook is configured
requests

# Parse configuration
pyyaml